// src/types.ts
export interface LoggerConfig {
	logFileName: string;
	fileLogLevel: string;
	consoleLogLevel: string;
}
// ...
export interface GenesisAssetEntry {
	module: string;
	data: Record<string, unknown>;
	schema: Schema;
}
// ...
export interface NetworkConfigLocal {
	name: string;
	tokenID: string;
	prevSnapshotBlockHeight: number;
	additionalAccounts: { address: Buffer; balance: bigint }[];
}